requirements.txt
mysql-connector-python==8.1.0
Faker==19.6.2